part 7 the importance of casting and arcitecture

+++ src/com/tutorial/main/GameObject.java
@@ -5,11 +5,11 @@ import java.awt.Rectangle;
 
 public abstract class GameObject {
 	
-	protected int x, y;
+	protected float x, y;
 	protected ID id;
-	protected int velX, velY;
+	protected float velX, velY;
 	
-	public GameObject(int x, int y, ID id) {
+	public GameObject(float x, float y, ID id) {
 		this.x = x;
 		this.y = y;
 		this.id = id;
@@ -34,19 +34,19 @@ public abstract class GameObject {
 	public void setVelY(int velY) {
 		this.velY = velY;
 	}
-	public int getX() {
+	public float getX() {
 		return x;
 	}
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 	public ID getID() {
 		return id;
 	}
-	public int getVelX() {
+	public float getVelX() {
 		return velX;
 	}
-	public int setVelY() {
+	public float setVelY() {
 		return velY;
 	}
 }
